refactor test use real ansi parser
